+++ docs/docs/quick_trade/plots.md
@@ -1,4 +1,4 @@
-# _plots
+# plots
 
 File for creating quick-trade charts.
