+++ quick_trade/plots.py
